+++ packages/core/src/combobox/combobox-base.tsx
@@ -30,7 +30,7 @@ import {
 } from "solid-js";
 
 import { createFormControl, FORM_CONTROL_PROP_NAMES, FormControlContext } from "../form-control";
-import { createFilter, createMessageFormatter } from "../i18n";
+import { createFilter } from "../i18n";
 import { createListState, ListKeyboardDelegate } from "../list";
 import { announce } from "../live-announcer";
 import { AsChildProp, Polymorphic } from "../polymorphic";
@@ -52,7 +52,7 @@ import {
   SelectionBehavior,
   SelectionMode,
 } from "../selection";
-import { COMBOBOX_INTL_MESSAGES } from "./combobox.intl";
+import { COMBOBOX_INTL_TRANSLATIONS, ComboboxIntlTranslations } from "./combobox.intl";
 import { ComboboxContext, ComboboxContextValue, ComboboxDataSet } from "./combobox-context";
 import { ComboboxTriggerMode } from "./types";
 
@@ -69,6 +69,9 @@ export interface ComboboxBaseSectionComponentProps<T> {
 export interface ComboboxBaseOptions<Option, OptGroup = never>
   extends Omit<PopperRootOptions, "anchorRef" | "contentRef" | "onCurrentPlacementChange">,
     AsChildProp {
+  /** The localized strings of the component. */
+  translations?: ComboboxIntlTranslations;
+
   /** The controlled open state of the combobox. */
   open?: boolean;
 
@@ -246,6 +249,7 @@ export function ComboboxBase<Option, OptGroup = never>(props: ComboboxBaseProps<
       preventScroll: false,
       defaultFilter: "contains",
       triggerMode: "input",
+      translations: COMBOBOX_INTL_TRANSLATIONS,
     },
     props,
   );
@@ -253,6 +257,7 @@ export function ComboboxBase<Option, OptGroup = never>(props: ComboboxBaseProps<
   const [local, popperProps, formControlProps, others] = splitProps(
     props,
     [
+      "translations",
       "itemComponent",
       "sectionComponent",
       "open",
@@ -317,8 +322,6 @@ export function ComboboxBase<Option, OptGroup = never>(props: ComboboxBaseProps<
   const [showAllOptions, setShowAllOptions] = createSignal(false);
 
   const [lastDisplayedOptions, setLastDisplayedOptions] = createSignal(local.options);
-
-  const messageFormatter = createMessageFormatter(() => COMBOBOX_INTL_MESSAGES);
 
   const disclosureState = createDisclosureState({
     open: () => local.open,
@@ -418,16 +421,23 @@ export function ComboboxBase<Option, OptGroup = never>(props: ComboboxBaseProps<
       return (local.options as Option[]).filter(filterFn);
     }
 
-    return local.options.filter(optGroup => {
+    const filteredGroups: OptGroup[] = [];
+    for (const optGroup of local.options as OptGroup[]) {
+      // Filter options of the group
       const filteredChildrenOptions = ((optGroup as any)[optionGroupChildren] as Option[]).filter(
         filterFn,
       );
+      // Don't add any groups that are empty
+      if (filteredChildrenOptions.length === 0) continue;
 
-      return {
+      // Add the group with the filtered options
+      filteredGroups.push({
         ...optGroup,
         [optionGroupChildren]: filteredChildrenOptions,
-      };
-    });
+      });
+    }
+
+    return filteredGroups;
   });
 
   const displayedOptions = createMemo(() => {
@@ -666,10 +676,8 @@ export function ComboboxBase<Option, OptGroup = never>(props: ComboboxBaseProps<
     if (isAppleDevice() && focusedItem != null && focusedKey !== lastAnnouncedFocusedKey) {
       const isSelected = listState.selectionManager().isSelected(focusedKey);
 
-      const announcement = messageFormatter().format("focusAnnouncement", {
-        optionText: focusedItem?.textValue || "",
-        isSelected,
-      });
+      const announcement =
+        local.translations?.focusAnnouncement(focusedItem?.textValue || "", isSelected) ?? "";
 
       announce(announcement);
     }
@@ -694,7 +702,7 @@ export function ComboboxBase<Option, OptGroup = never>(props: ComboboxBaseProps<
       isOpen !== lastOpen && (listState.selectionManager().focusedKey() == null || isAppleDevice());
 
     if (isOpen && (didOpenWithoutFocusedItem || optionCount !== lastOptionCount)) {
-      const announcement = messageFormatter().format("countAnnouncement", { optionCount });
+      const announcement = local.translations?.countAnnouncement(optionCount) ?? "";
       announce(announcement);
     }
 
@@ -716,9 +724,8 @@ export function ComboboxBase<Option, OptGroup = never>(props: ComboboxBaseProps<
       lastSelectedItem &&
       lastSelectedKey !== lastAnnouncedSelectedKey
     ) {
-      const announcement = messageFormatter().format("selectedAnnouncement", {
-        optionText: lastSelectedItem?.textValue || "",
-      });
+      const announcement =
+        local.translations?.selectedAnnouncement(lastSelectedItem?.textValue || "") ?? "";
 
       announce(announcement);
     }
@@ -758,8 +765,8 @@ export function ComboboxBase<Option, OptGroup = never>(props: ComboboxBaseProps<
     listState: () => listState,
     keyboardDelegate: delegate,
     listboxId,
-    triggerAriaLabel: () => messageFormatter().format("triggerLabel"),
-    listboxAriaLabel: () => messageFormatter().format("listboxLabel"),
+    triggerAriaLabel: () => local.translations?.triggerLabel,
+    listboxAriaLabel: () => local.translations?.listboxLabel,
     setIsInputFocused,
     resetInputValue,
     setInputValue,