+++ packages/core/src/menu/menu-trigger.tsx
@@ -7,11 +7,12 @@
  */
 
 import { callHandler, mergeDefaultProps, mergeRefs, OverrideComponentProps } from "@kobalte/utils";
-import { createEffect, JSX, onCleanup, splitProps } from "solid-js";
+import { createDeferred, createEffect, createSignal, JSX, onCleanup, splitProps } from "solid-js";
 
 import * as Button from "../button";
 import { useMenuContext } from "./menu-context";
 import { useMenuRootContext } from "./menu-root-context";
+import { useOptionalMenubarContext } from "../menubar/menubar-context";
 
 export interface MenuTriggerOptions extends Button.ButtonRootOptions {}
 
@@ -21,14 +22,17 @@ export interface MenuTriggerProps extends OverrideComponentProps<"button", MenuT
  * The button that toggles the menu.
  */
 export function MenuTrigger(props: MenuTriggerProps) {
+  let ref: HTMLButtonElement | undefined;
+
   const rootContext = useMenuRootContext();
   const context = useMenuContext();
+  const optionalMenubarContext = useOptionalMenubarContext();
 
   props = mergeDefaultProps(
     {
       id: rootContext.generateId("trigger"),
     },
-    props,
+    props
   );
 
   const [local, others] = splitProps(props, [
@@ -38,7 +42,39 @@ export function MenuTrigger(props: MenuTriggerProps) {
     "onPointerDown",
     "onClick",
     "onKeyDown",
+    "onMouseOver",
+    "onFocus",
   ]);
+
+  let key: string | undefined;
+
+  if (optionalMenubarContext !== undefined) {
+    key = rootContext.value() ?? local.id!;
+
+    createEffect(() => {
+      optionalMenubarContext.registerMenu(key!, [
+        context.contentRef() ?? ref!,
+        ...context.nestedMenus(),
+      ]);
+    });
+
+    createEffect(() => {
+      if (optionalMenubarContext.value() === key) {
+        ref?.focus();
+        if (optionalMenubarContext.autoFocusMenu()) context.open(true);
+      } else context.close(true);
+    });
+
+    createEffect(() => {
+      if (context.isOpen()) optionalMenubarContext.setValue(key);
+    });
+
+    onCleanup(() => {
+      optionalMenubarContext.unregisterMenu(key!);
+    });
+
+    if (optionalMenubarContext.lastValue() === undefined) optionalMenubarContext.setLastValue(key);
+  }
 
   const onPointerDown: JSX.EventHandlerUnion<any, PointerEvent> = e => {
     callHandler(e, local.onPointerDown);
@@ -47,15 +83,20 @@ export function MenuTrigger(props: MenuTriggerProps) {
 
     // For consistency with native, open the select on mouse down (main button), but touch up.
     if (!local.disabled && e.pointerType !== "touch" && e.button === 0) {
-      context.toggle(true);
+      // When opened by click, automatically focus Menubar menus
+      optionalMenubarContext?.setAutoFocusMenu(true);
+
+      // Don't auto focus element for Menubar
+      if (optionalMenubarContext !== undefined) context.toggle(false);
+      else context.toggle(true);
     }
   };
 
   const onClick: JSX.EventHandlerUnion<any, MouseEvent> = e => {
     callHandler(e, local.onClick);
 
-    if (!local.disabled && e.currentTarget.dataset.pointerType === "touch") {
-      context.toggle(true);
+    if (!local.disabled) {
+      if (e.currentTarget.dataset.pointerType === "touch") context.toggle(true);
     }
   };
 
@@ -80,22 +121,68 @@ export function MenuTrigger(props: MenuTriggerProps) {
         e.preventDefault();
         context.toggle("last");
         break;
+      case "ArrowRight":
+        if (optionalMenubarContext === undefined) break;
+        e.stopPropagation();
+        e.preventDefault();
+        optionalMenubarContext.nextMenu();
+        break;
+      case "ArrowLeft":
+        if (optionalMenubarContext === undefined) break;
+        e.stopPropagation();
+        e.preventDefault();
+        optionalMenubarContext.previousMenu();
+        break;
     }
+  };
+
+  const onMouseOver: JSX.EventHandlerUnion<any, MouseEvent> = e => {
+    callHandler(e, local.onMouseOver);
+
+    // When one of the menubar menus is open, automatically open others on trigger hover
+    if (
+      !local.disabled &&
+      optionalMenubarContext !== undefined &&
+      optionalMenubarContext.value() !== undefined
+    ) {
+      optionalMenubarContext.setValue(key);
+    }
+  };
+
+  const onFocus: JSX.EventHandlerUnion<any, FocusEvent> = e => {
+    callHandler(e, local.onFocus);
+
+    if (optionalMenubarContext !== undefined) optionalMenubarContext.setValue(key);
   };
 
   createEffect(() => onCleanup(context.registerTriggerId(local.id!)));
 
+  const [tabIndex, setTabIndex] = createSignal<undefined | 0 | -1>(undefined);
+
   return (
     <Button.Root
-      ref={mergeRefs(context.setTriggerRef, local.ref)}
+      ref={mergeRefs(el => (ref = el), context.setTriggerRef, local.ref)}
       id={local.id}
       disabled={local.disabled}
       aria-haspopup="true"
       aria-expanded={context.isOpen()}
       aria-controls={context.isOpen() ? context.contentId() : undefined}
+      data-highlighted={
+        key !== undefined && optionalMenubarContext?.value() === key ? true : undefined
+      }
+      tabIndex={
+        optionalMenubarContext !== undefined
+          ? optionalMenubarContext.value() === key || optionalMenubarContext.lastValue() === key
+            ? 0
+            : -1
+          : undefined
+      }
       onPointerDown={onPointerDown}
+      onMouseOver={onMouseOver}
       onClick={onClick}
       onKeyDown={onKeyDown}
+      onFocus={onFocus}
+      role={optionalMenubarContext !== undefined ? "menuitem" : undefined}
       {...context.dataset()}
       {...others}
     />