+++ packages/core/src/slider/utils.ts
@@ -16,7 +16,9 @@ export function getClosestValueIndex(values: number[], nextValue: number) {
   if (values.length === 1) return 0;
   const distances = values.map(value => Math.abs(value - nextValue));
   const closestDistance = Math.min(...distances);
-  return distances.indexOf(closestDistance);
+  const closestIndex = distances.indexOf(closestDistance);
+
+  return nextValue < values[closestIndex] ? closestIndex : distances.lastIndexOf(closestDistance);
 }
 
 /**
@@ -67,4 +69,9 @@ export function linearScale(input: readonly [number, number], output: readonly [
     const ratio = (output[1] - output[0]) / (input[1] - input[0]);
     return output[0] + ratio * (value - input[0]);
   };
+}
+
+export function stopEventDefaultAndPropagation(event: Event) {
+  event.preventDefault();
+  event.stopPropagation();
 }