+++ packages/dates/src/date-picker/date-picker-root.tsx
@@ -19,7 +19,6 @@ import {
   createDisclosureState,
   createFormControl,
   createFormResetListener,
-  createMessageFormatter,
   createPresence,
   createRegisterId,
   FORM_CONTROL_PROP_NAMES,
@@ -64,7 +63,7 @@ import {
   getFirstValueOfSelection,
   isDateInvalid,
 } from "../calendar/utils";
-import { DATE_PICKER_INTL_MESSAGES } from "./date-picker.intl";
+import { DATE_PICKER_INTL_TRANSLATIONS, DatePickerIntlTranslations } from "./date-picker.intl";
 import {
   DatePickerContext,
   DatePickerContextValue,
@@ -201,6 +200,9 @@ export type DatePickerRootOptions = (
     /** Whether the date picker is read only. */
     readOnly?: boolean;
 
+    /** The localized strings of the component. */
+    translations?: DatePickerIntlTranslations;
+
     /** The children of the date picker. */
     children?: JSX.Element;
   };
@@ -225,6 +227,7 @@ export function DatePickerRoot(props: DatePickerRootProps) {
       gutter: 8,
       sameWidth: false,
       placement: "bottom-start",
+      translations: DATE_PICKER_INTL_TRANSLATIONS,
     },
     props,
   );
@@ -232,6 +235,7 @@ export function DatePickerRoot(props: DatePickerRootProps) {
   const [local, popperProps, formControlProps, others] = splitProps(
     props,
     [
+      "translations",
       "locale",
       "createCalendar",
       "visibleDuration",
@@ -281,8 +285,6 @@ export function DatePickerRoot(props: DatePickerRootProps) {
   const [controlRef, setControlRef] = createSignal<HTMLDivElement>();
   const [triggerRef, setTriggerRef] = createSignal<HTMLButtonElement>();
   const [contentRef, setContentRef] = createSignal<HTMLDivElement>();
-
-  const messageFormatter = createMessageFormatter(() => DATE_PICKER_INTL_MESSAGES);
 
   const locale = createMemo(() => {
     return local.locale ?? useLocale().locale();
@@ -405,9 +407,7 @@ export function DatePickerRoot(props: DatePickerRootProps) {
     let description = "";
 
     if (local.selectionMode === "single" || local.selectionMode === "multiple") {
-      description = messageFormatter().format("selectedDateDescription", {
-        date: formattedValue(),
-      });
+      description = local.translations?.selectedDateDescription(formattedValue()) ?? "";
     } else if (local.selectionMode === "range") {
       // TODO: RangeDatePicker
     }
@@ -522,12 +522,12 @@ export function DatePickerRoot(props: DatePickerRootProps) {
   );
 
   const context: DatePickerContextValue = {
+    translations: () => local.translations ?? DATE_PICKER_INTL_TRANSLATIONS,
     dataset,
     isOpen: disclosureState.isOpen,
     isDisabled: () => formControlContext.isDisabled() ?? false,
     isModal: () => local.modal ?? false,
     contentPresence,
-    messageFormatter,
     granularity,
     maxGranularity: () => local.maxGranularity,
     hourCycle: () => local.hourCycle,